refactor: 폴더 네이밍 컨벤션 복수형으로 통일
- view → views, viewmodel → viewmodels, service → services
- 관련 import 경로 일괄 수정
- CLAUDE.md 프로젝트 구조 문서 업데이트
