+++ lib/core/config/provider_config.dart
@@ -3,16 +3,16 @@ import 'package:arttrip/core/network/network.dart';
 import 'package:arttrip/features/exhibit/data/exhibit_repository.dart';
 import 'package:arttrip/features/exhibit/data/exhibit_repository_hybrid.dart';
 import 'package:arttrip/features/exhibit/data/exhibit_repository_mock.dart';
-import 'package:arttrip/features/exhibit/viewmodel/exhibit_detail_viewmodel.dart';
-import 'package:arttrip/features/exhibit/viewmodel/exhibit_viewmodel.dart';
-import 'package:arttrip/features/exhibit/viewmodel/write_review_viewmodel.dart';
+import 'package:arttrip/features/exhibit/viewmodels/exhibit_detail_viewmodel.dart';
+import 'package:arttrip/features/exhibit/viewmodels/exhibit_viewmodel.dart';
+import 'package:arttrip/features/exhibit/viewmodels/write_review_viewmodel.dart';
 import 'package:arttrip/features/home/home_repository.dart';
 import 'package:arttrip/features/home/home_repository_hybrid.dart';
 import 'package:arttrip/features/home/home_repository_mock.dart';
 import 'package:arttrip/features/home/home_viewmodel.dart';
 import 'package:arttrip/features/onboarding/data/keywords_repository.dart';
 import 'package:arttrip/features/onboarding/data/keywords_repository_mock.dart';
-import 'package:arttrip/features/onboarding/viewmodel/keywords_viewmodel.dart';
+import 'package:arttrip/features/onboarding/viewmodels/keywords_viewmodel.dart';
 import 'package:arttrip/shared/viewmodels/alert_viewmodel.dart';
 import 'package:provider/provider.dart';
