+++ lib/features/exhibit/viewmodels/exhibit_detail_viewmodel.dart
