+++ lib/features/exhibit/widgets/write_review/review_content_section.dart
@@ -1,6 +1,6 @@
 import 'package:arttrip/core/app_colors.dart';
 import 'package:arttrip/core/extensions.dart';
-import 'package:arttrip/features/exhibit/viewmodel/write_review_viewmodel.dart';
+import 'package:arttrip/features/exhibit/viewmodels/write_review_viewmodel.dart';
 import 'package:arttrip/shared/utils/text/arttrip_text.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
