+++ lib/features/home/views/today_exhibits_recommendation_view.dart
@@ -1,7 +1,7 @@
 import 'package:arttrip/core/app_consts.dart';
 import 'package:arttrip/core/extensions.dart';
 import 'package:arttrip/features/exhibit/data/models/exhibit_model.dart';
-import 'package:arttrip/features/exhibit/viewmodel/exhibit_viewmodel.dart';
+import 'package:arttrip/features/exhibit/viewmodels/exhibit_viewmodel.dart';
 import 'package:arttrip/features/home/home_viewmodel.dart';
 import 'package:arttrip/features/home/widgets/today_exhibit_widget.dart';
 import 'package:arttrip/routes/routes.dart';
