+++ lib/features/splash/views/splash_view.dart
@@ -1,5 +1,5 @@
 import 'package:arttrip/core/app_colors.dart';
-import 'package:arttrip/features/auth/service/auth_service.dart';
+import 'package:arttrip/features/auth/services/auth_service.dart';
 import 'package:arttrip/routes/routes.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_native_splash/flutter_native_splash.dart';
