+++ lib/features/stamp/views/stamp_view.dart
