+++ lib/main.dart
@@ -4,8 +4,8 @@ import 'package:arttrip/core/config/provider_config.dart';
 import 'package:arttrip/core/enum.dart';
 import 'package:arttrip/core/env.dart';
 import 'package:arttrip/core/network/network.dart';
-import 'package:arttrip/features/auth/service/auth_service.dart';
-import 'package:arttrip/features/auth/service/token_storage_service.dart';
+import 'package:arttrip/features/auth/services/auth_service.dart';
+import 'package:arttrip/features/auth/services/token_storage_service.dart';
 import 'package:arttrip/l10n/generated/app_localizations.dart';
 import 'package:arttrip/routes/app_router.dart';
 import 'package:flutter/material.dart';
